+++ Assets/Scripts/Services/Input/Processors.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 96e8c7f0c2c352148bba9214f61cf1d3
+guid: a1060b11857448b883f2187f26550e8d
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
