+++ Assets/Scripts/Services/Input/Processors/RelativeToScreenVector2Processor.cs
@@ -1,35 +1,34 @@
-﻿using UnityEditor;
+﻿#if UNITY_EDITOR
+using UnityEditor;
+#endif
 using UnityEngine;
 using UnityEngine.InputSystem;
 
 namespace Services.Input
 {
+
 #if UNITY_EDITOR
     [InitializeOnLoad]
 #endif
-    public class OffsetVector2Processor : InputProcessor<Vector2>
+    public class RelativeToScreenVector2Processor : InputProcessor<Vector2>
     {
 
 #if UNITY_EDITOR
-        static OffsetVector2Processor()
+        static RelativeToScreenVector2Processor()
         {
             Initialize();
         }
 #endif
 
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.BeforeSceneLoad)]
-        static void Initialize()
+        static void Initialize()    
         {
-            InputSystem.RegisterProcessor<OffsetVector2Processor>();
+            InputSystem.RegisterProcessor<RelativeToScreenVector2Processor>();
         }
-
-        [Tooltip("Vector to add to incoming values.")]
-        public float X = 0;
-        public float Y = 0;
 
         public override Vector2 Process(Vector2 value, InputControl control)
         {
-            return new Vector2(value.x + X, value.y + Y);
+            return new Vector2(value.x / Screen.width, value.y / Screen.height);
         }
     }
 }
